liveness-server/requirements.txt
fastapi==0.115.6
uvicorn==0.33.0
opencv-python==4.10.0.84
mediapipe==0.10.11
pillow==10.4.0
numpy==1.24.4
protobuf<4
